# Performance improvement: record-batching