--- a/.vimrc
+++ b/.vimrc
@@ -1,1 +1,2 @@
 map \d :w<C-m>:!clear;echo Building ...; echo; make build<C-m>
+map \f :w<C-m>:!clear;echo Building ...; echo; make mall<C-m>