--- a/Makefile
+++ b/Makefile
@@ -77,6 +77,11 @@ mprof:
 	go build github.com/johnkerl/miller/cmd/mprof
 mprof2:
 	go build github.com/johnkerl/miller/cmd/mprof2
+mprof3:
+	go build github.com/johnkerl/miller/cmd/mprof3
+mprof4:
+	go build github.com/johnkerl/miller/cmd/mprof4
+mall: mprof4 mprof3 mprof2 mprof mlr
 
 # Please see comments in ./create-release-tarball as well as
 # https://miller.readthedocs.io/en/latest/build/#creating-a-new-release-for-developers
@@ -84,4 +89,4 @@ release_tarball: build check
 	./create-release-tarball
 
 # Go does its own dependency management, outside of make.
-.PHONY: build mlr mprof mprof2 check unit_test regression_test fmt dev
+.PHONY: build mlr mprof mprof2 mprof3 mprof4 check unit_test regression_test fmt dev