--- a/Makefile
+++ b/Makefile
@@ -73,6 +73,8 @@ so: install
 sure: build check
 mlr:
 	go build github.com/johnkerl/miller/cmd/mlr
+mprof:
+	go build github.com/johnkerl/miller/cmd/mprof
 
 # Please see comments in ./create-release-tarball as well as
 # https://miller.readthedocs.io/en/latest/build/#creating-a-new-release-for-developers
@@ -80,4 +82,4 @@ release_tarball: build check
 	./create-release-tarball
 
 # Go does its own dependency management, outside of make.
-.PHONY: build mlr check unit_test regression_test fmt dev
+.PHONY: build mlr mprof check unit_test regression_test fmt dev