--- a/Makefile
+++ b/Makefile
@@ -75,6 +75,8 @@ mlr:
 	go build github.com/johnkerl/miller/cmd/mlr
 mprof:
 	go build github.com/johnkerl/miller/cmd/mprof
+mprof2:
+	go build github.com/johnkerl/miller/cmd/mprof2
 
 # Please see comments in ./create-release-tarball as well as
 # https://miller.readthedocs.io/en/latest/build/#creating-a-new-release-for-developers
@@ -82,4 +84,4 @@ release_tarball: build check
 	./create-release-tarball
 
 # Go does its own dependency management, outside of make.
-.PHONY: build mlr mprof check unit_test regression_test fmt dev
+.PHONY: build mlr mprof mprof2 check unit_test regression_test fmt dev