--- a/cmd/mlr/main.go
+++ b/cmd/mlr/main.go
@@ -9,6 +9,7 @@ import (
 	"runtime/pprof"
 	"strconv"
 
+	"github.com/pkg/profile" // for trace.out
 	"github.com/johnkerl/miller/internal/pkg/entrypoint"
 )
 
@@ -59,7 +60,12 @@ func main() {
 		defer pprof.StopCPUProfile()
 
 		fmt.Fprintf(os.Stderr, "CPU profile started.\n")
-		defer fmt.Fprintf(os.Stderr, "CPU profile finished.\n")
+		defer fmt.Fprintf(os.Stderr, "CPU profile finished: go tool pprof -http=:8080 %s\n", profFilename)
+	}
+
+	if len(os.Args) >= 3 && os.Args[1] == "--traceprofile" {
+		defer profile.Start(profile.TraceProfile, profile.ProfilePath(".")).Stop()
+		defer fmt.Fprintf(os.Stderr, "go tool trace trace.out\n")
 	}
 
 	// This will obtain os.Args and go from there.  All the usual contents of