--- a/go.mod
+++ b/go.mod
@@ -1,4 +1,5 @@
 module github.com/johnkerl/miller
+
 // The repo is 'miller' and the executable is 'mlr', going back many years and
 // predating the Go port.
 //
@@ -16,6 +17,7 @@ require (
 	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51
 	github.com/lestrrat-go/strftime v1.0.4
 	github.com/mattn/go-isatty v0.0.12
+	github.com/pkg/profile v1.6.0 // indirect
 	golang.org/x/sys v0.0.0-20210326220804-49726bf1d181
 	golang.org/x/term v0.0.0-20201210144234-2321bbc49cbf
 )