--- a/scripts/goslow.sh
+++ b/scripts/goslow.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+cat "$@" | while read line; do
+  millisleep 100
+  echo $line
+done