… cmd/mprof3/main.go
… cmd/mprof5/main.go
… go.sum
… internal/pkg/auxents/repl/session.go
… internal/pkg/auxents/repl/types.go
… internal/pkg/auxents/repl/verbs.go
… internal/pkg/cli/option_parse.go
… internal/pkg/cli/option_types.go
… internal/pkg/climain/mlrcli_parse.go
… internal/pkg/dsl/cst/dump.go
… internal/pkg/dsl/cst/emit1.go
… internal/pkg/dsl/cst/emit_emitp.go
… internal/pkg/dsl/cst/emitf.go
… internal/pkg/dsl/cst/print.go
… internal/pkg/input/pseudo_reader_gen.go
… internal/pkg/input/record_reader.go
… internal/pkg/input/record_reader_csv.go
… internal/pkg/input/record_reader_csvlite.go
… internal/pkg/input/record_reader_dkvp.go
… internal/pkg/input/record_reader_dkvp_nidx.go
… internal/pkg/input/record_reader_factory.go
… internal/pkg/input/record_reader_json.go
… internal/pkg/input/record_reader_nidx.go
… internal/pkg/input/record_reader_xtab.go
… internal/pkg/output/channel_writer.go
… internal/pkg/output/file-output-handlers.go
… internal/pkg/runtime/state.go
… internal/pkg/stream/stream.go
… internal/pkg/transformers/aaa_chain_transformer.go
… internal/pkg/transformers/aaa_record_transformer.go