This is a second performance PR, following after https://github.com/johnkerl/miller/pull/774.

This affects all [Miller-supported file formats](https://miller.readthedocs.io/en/latest/file-formats). Some performance numbers for CSV, on a Mac laptop:

* Million-line CSV input which is an expansion of https://github.com/johnkerl/miller/blob/main/docs/src/example.csv
* Miller 5 `cat`: 2.423 seconds
* Miller 6 `cat` before this PR: 4.175 seconds
* Miller 6 `cat` with this PR: 3.106 seconds
* Miller 6 `cat` with this PR and `mlr -S`: 1.873 seconds

Note that the last number is the runtime we'll have after the WIP [JIT branch](https://github.com/johnkerl/miller/tree/jit-type-infer) is PR'ed and merged. At that point, single-verb performance will be noticeably better than Miller 5. Furthermore, for multi-verb then-chains, the performance improvement will be even more noticeable, since Miller <= 5 (in C) is single-core whereas Miller 6 (in Go) is multicore.

See also new files in the `scripts` directory on this PR which help for doing measurements of this sort.